fix: Move `refreshConnectionStatuses` to separate bean
Move the implementation of the `refreshConnectionStatuses` functionality to a separate bean, which hopefully
fixes some flaky tests that might be caused by `@InjectMock`ing a bean that defines methods with `@Scheduled`
annotations.
