+++ src/main/java/io/confluent/idesidecar/restapi/connections/ConnectionStateManager.java
@@ -14,7 +14,6 @@ import io.confluent.idesidecar.restapi.exceptions.Failure.Error;
 import io.confluent.idesidecar.restapi.exceptions.InvalidInputException;
 import io.confluent.idesidecar.restapi.models.ConnectionSpec;
 import io.confluent.idesidecar.restapi.util.UuidFactory;
-import io.quarkus.scheduler.Scheduled;
 import io.smallrye.mutiny.Uni;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.event.Event;
@@ -23,7 +22,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.function.Predicate;
 
 /**
  * The {@link ConnectionStateManager} allows to manage a set of {@link ConnectionState}s. It
@@ -294,37 +292,5 @@ public class ConnectionStateManager {
   // This method should be only used in tests.
   public void clearAllConnectionStates() {
     connectionStates.clear();
-  }
-
-  @Scheduled(
-      every = "${ide-sidecar.connections.ccloud.refresh-status-interval-seconds}s"
-  )
-  void refreshConfluentCloudConnectionStatuses() {
-    refreshConnectionStatuses(CCloudConnectionState.class::isInstance);
-  }
-
-  @Scheduled(
-      every = "${ide-sidecar.connections.confluent-local.refresh-status-interval-seconds}s"
-  )
-  void refreshLocalConnectionStatuses() {
-    refreshConnectionStatuses(LocalConnectionState.class::isInstance);
-  }
-
-  @Scheduled(
-      every = "${ide-sidecar.connections.direct.refresh-status-interval-seconds}s"
-  )
-  void refreshDirectConnectionStatuses() {
-    refreshConnectionStatuses(DirectConnectionState.class::isInstance);
-  }
-
-  /**
-   * Refreshes the statuses of all connections that match a given predicate.
-   *
-   * @param predicate The predicate that connections must match to be refreshed.
-   */
-  void refreshConnectionStatuses(Predicate<ConnectionState> predicate) {
-    getConnectionStates().stream()
-        .filter(predicate)
-        .forEach(ConnectionState::refreshStatus);
   }
 }
