+++ src/test/java/io/confluent/idesidecar/restapi/auth/RefreshCCloudTokensBeanTest.java
@@ -8,27 +8,18 @@ import io.confluent.idesidecar.restapi.connections.LocalConnectionState;
 import io.confluent.idesidecar.restapi.connections.PlatformConnectionState;
 import io.confluent.idesidecar.restapi.models.ConnectionSpec;
 import io.confluent.idesidecar.restapi.models.ConnectionSpec.ConnectionType;
-import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
 import io.vertx.core.Future;
-import jakarta.inject.Inject;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
-import org.junitpioneer.jupiter.RetryingTest;
 import org.mockito.Mockito;
 
 @QuarkusTest
 class RefreshCCloudTokensBeanTest {
-  @InjectMock
-  ConnectionStateManager connectionStateManager;
-
-  @Inject
-  RefreshCCloudTokensBean refreshCCloudTokensBean;
-
-  @RetryingTest(maxAttempts = 5)
+  @Test
   void refreshTokensShouldConsiderOnlyCCloudConnections() {
     var confluentCloudConnection = (CCloudConnectionState) createdSpiedConnectionState(
         "1", "1", ConnectionType.CCLOUD);
@@ -37,7 +28,10 @@ class RefreshCCloudTokensBeanTest {
         confluentCloudConnection,
         localConnection
     );
+    var connectionStateManager = Mockito.spy(ConnectionStateManager.class);
     Mockito.when(connectionStateManager.getConnectionStates()).thenReturn(connections);
+
+    var refreshCCloudTokensBean = new RefreshCCloudTokensBean(connectionStateManager);
 
     refreshCCloudTokensBean.refreshTokens();
 
@@ -45,7 +39,7 @@ class RefreshCCloudTokensBeanTest {
   }
 
   @DisabledIfSystemProperty(named = "os.name", matches = ".*Windows.*")
-  @RetryingTest(maxAttempts = 5)
+  @Test
   void refreshTokensShouldRefreshOnlyConnectionsEligibleForATokenRefreshAttempt() {
     // Connection eligible for a token refresh attempt
     var eligibleConnection = (CCloudConnectionState) createdSpiedConnectionState(
@@ -79,7 +73,10 @@ class RefreshCCloudTokensBeanTest {
         "local", "name", ConnectionType.LOCAL);
 
     var connections = List.of(eligibleConnection, ineligibleConnection, localConnection);
+    var connectionStateManager = Mockito.spy(ConnectionStateManager.class);
     Mockito.when(connectionStateManager.getConnectionStates()).thenReturn(connections);
+
+    var refreshCCloudTokensBean = new RefreshCCloudTokensBean(connectionStateManager);
 
     refreshCCloudTokensBean.refreshTokens();
